pyproject.toml
[project]
name = "hanzi-similar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.20",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.1",
    "fonttools>=4.59.1",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "scikit-learn>=1.7.1",
    "sentence-transformers>=5.1.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.55.3",
    "uvicorn>=0.35.0",
]
